ユニークな文字列の生成をモジュールに分割
